LCU and Helpers now follow the strucutre
