+++ src-tauri/src/main.rs
@@ -10,7 +10,7 @@ use structs::{Shards, Data, ItemsMap, AbilitiesMap};
 use constants::{TIERS, REGIONS, ROLES};
 use frontend_types::ChampionInfo;
 
-use shared::helpers::ChampionNames;
+use shared::helpers::structs::ChampionNames;
 
 use crate::frontend_types::RuneImages;
 
@@ -71,7 +71,7 @@ async fn rune_names(
 
 #[tauri::command]
 async fn champion_names() -> Result<Vec<ChampionNames>, i64> {
-    let request = shared::helpers::all_champion_names().await;
+    let request = shared::helpers::champs::all_champion_names().await;
     match request {
         Ok(names) => Ok(names),
         Err(err) => Err(err),
