+++ src-tauri/src/plugins/ugg/runes.rs
@@ -22,7 +22,7 @@ impl structs::Data {
                     return Err(206);
                 };
 
-                let all_runes = helpers::all_rune_images(*tree_id_one, *tree_id_two, "es_ES".to_string()).await;
+                let all_runes = helpers::runes::all_rune_images(*tree_id_one, *tree_id_two, "es_ES".to_string()).await;
                 match all_runes {
                     Ok(immutable_all_runes) => {
                         let mut used_rune_ids = Vec::new();
